size_t fixes for building on win x64
Fixes C4267 warnings, and errors regarding passing size_t* when they chould be unsigned int *.

winrt_utils:
WindowsCreateString takes a UINT32 as size. Truncation would mean
shorter than expected string, but presumably not a severe problem
in this case.

file_picker:
The interface uses unsigned int, not size_t, so change implementation and callers to unsigned int.


R=cpu@chromium.org
BUG=166496


Review URL: https://chromiumcodereview.appspot.com/11884021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@176854 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d67eb0ff81a18335f099763db733cec9756c5b93

+++ win8/metro_driver/winrt_utils.cc
@@ -23,7 +23,8 @@ void CheckHR(HRESULT hr, const char* message) {
 
 HSTRING MakeHString(const string16& str) {
   HSTRING hstr;
-  if (FAILED(::WindowsCreateString(str.c_str(), str.size(), &hstr))) {
+  if (FAILED(::WindowsCreateString(str.c_str(), static_cast<UINT32>(str.size()),
+                                   &hstr))) {
     PLOG(DFATAL) << "Hstring creation failed";
   }
   return hstr;
